TFormula in SignalShapeService made ROOT6 friendly
ROOT 6 TFormula reacts badly if a parameter is implied, as the 4th
parameter of the induction plane frequency filter was.
In that case, values were assigned to the wrong parameters, resulting in
nonsensical results.

Another change that I am not sure is necessary in this case: replaced
"(x > 0.)*expr(x)" with a ternary operator "(x<=0.)? 0.0: expr(x)".
If expr(x) on a non-positive x evaluates to NaN or infinity, the result
was be NaN, now it is 0 (in other words, the expression was always
evaluated and in special cases it could matter).